tools/osuosz to bms (modified om2bms)/requirements.txt
pillow
pydub